move inchikeyPairGenerator to data_generators.py
