+++ ms2deepscore/train_new_model/__init__.py
@@ -1,10 +1,8 @@
-from .data_generators import DataGeneratorPytorch
-from .inchikey_pair_selection import (InchikeyPairGenerator,
-                                      select_compound_pairs_wrapper)
+from .data_generators import DataGeneratorPytorch, InchikeyPairGenerator
+from .inchikey_pair_selection import (select_compound_pairs_wrapper)
 
 
 __all__ = [
     "DataGeneratorPytorch",
     "select_compound_pairs_wrapper",
-    "InchikeyPairGenerator",
 ]
