+++ tests/test_data_generators.py
@@ -7,8 +7,7 @@ from ms2deepscore.SettingsMS2Deepscore import SettingsMS2Deepscore, SettingsEmbe
 from ms2deepscore.tensorize_spectra import tensorize_spectra
 from ms2deepscore.train_new_model.data_generators import DataGeneratorPytorch, \
     DataGeneratorEmbeddingEvaluation, create_data_generator
-from ms2deepscore.train_new_model.inchikey_pair_selection import \
-    InchikeyPairGenerator
+from ms2deepscore.train_new_model import InchikeyPairGenerator
 from tests.create_test_spectra import create_test_spectra
